authservice/authapp/requirements.txt
djangorestframework
requests
python-dotenv
djangorestframework-simplejwt
psycopg[binary,pool]
pyotp

#for testing mfa qr
Pillow
qrcode